feat: added HarmonicCache
Each harmonic and its derivatives are evaluated 4 times at the same
coordinates, and many values stay the same, so it makes sense to cache
them, since they are quite expensive. Similar to the Accelerators, they
are stored inside State, and do not affect the behavior of the
equilibrium objects themselves.

updated doc examples
updated tests